+++ src/test/java/org/opencloudb/performance/TestUpdatePerf.java
@@ -23,7 +23,7 @@
  */
 package org.opencloudb.performance;
 
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 /**
  * @author wuzh
@@ -51,9 +51,9 @@ public class TestUpdatePerf extends AbstractMultiTreadBatchTester {
 	}
 
 	@Override
-	public Runnable createJob(SimpleConPool conPool2, int myCount, int batch,
-			int startId, AtomicInteger finshiedCount2,
-			AtomicInteger failedCount2) {
+	public Runnable createJob(SimpleConPool conPool2, long myCount, int batch,
+			long startId, AtomicLong finshiedCount2,
+			AtomicLong failedCount2) {
 		return new TravelRecordUpdateJob(conPool2, myCount, batch, startId,
 				finshiedCount, failedCount);
 	}
